Refactoring total funding fetching on <DeployedParams>
